+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # PyGPSClient Release Notes
 
+### RELEASE 1.5.2
+
+1. Fix logpath setting in config file - Fixes [#171](https://github.com/semuconsulting/PyGPSClient/issues/171)
+1. Add support for pyubxutils.
+1. Minor updates to vscode task configurations.
+
 ### RELEASE 1.5.1
 
 1. Add new user-selectable and configurable "Chart" widget: